+++ stroom-security-server/src/main/java/stroom/security/server/JWTService.java
@@ -43,6 +43,7 @@ public class JWTService implements HasHealthCheck {
     private PublicJsonWebKey jwk;
     private Duration durationToWarnBeforeExpiry = null;
     private final String apiKey;
+    private final String authenticationServiceUrl;
     private final String authJwtIssuer;
     private AuthenticationServiceClients authenticationServiceClients;
     private final boolean checkTokenRevocation;
@@ -62,6 +63,7 @@ public class JWTService implements HasHealthCheck {
             this.durationToWarnBeforeExpiry = Duration.ofMillis(ModelStringUtil.parseDurationString(durationToWarnBeforeExpiry));
         }
         this.apiKey = apiKey;
+        this.authenticationServiceUrl = authenticationServiceUrl;
         this.authJwtIssuer = authJwtIssuer;
         this.authenticationServiceClients = authenticationServiceClients;
         this.checkTokenRevocation = enableTokenRevocationCheck;
@@ -99,10 +101,17 @@ public class JWTService implements HasHealthCheck {
      * We need to do this if the remote public key changes and verification fails.
      */
     private String fetchNewPublicKey() throws ApiException {
-        // We need to fetch the public key from the remote authentication service.
-        final ApiKeyApi apiKeyApi = authenticationServiceClients.newApiKeyApi();
-        String jwkAsJson = apiKeyApi.getPublicKey();
-        return jwkAsJson;
+        try {
+            // We need to fetch the public key from the remote authentication service.
+            final ApiKeyApi apiKeyApi = authenticationServiceClients.newApiKeyApi();
+            return apiKeyApi.getPublicKey();
+        } catch (final ApiException | RuntimeException e) {
+            LOGGER.error("Error fetching new public API key from URL '" + authenticationServiceUrl + "'.");
+            if (authenticationServiceUrl != null && authenticationServiceUrl.toLowerCase().contains("https")) {
+                LOGGER.error("Are you sure you want to use HTTPS? IF so you will need to configure the trust store or disable SSL verification with 'stroom.auth.services.verifyingSsl=false'");
+            }
+            throw e;
+        }
     }
 
     public boolean containsValidJws(ServletRequest request) {
@@ -151,7 +160,7 @@ public class JWTService implements HasHealthCheck {
         } catch (InvalidJwtException e) {
             try {
                 JwtClaims claims = e.getJwtContext().getJwtClaims();
-                if(claims.getExpirationTime().getValueInMillis() < Instant.now().toEpochMilli() ){
+                if (claims.getExpirationTime().getValueInMillis() < Instant.now().toEpochMilli()) {
                     LOGGER.info("The API key for user '{}' has expired. An API key is required, i.e. for queries. Creating a new one.", claims.getSubject());
                     String newJws = authenticationServiceClients.createTokenForUser(claims.getSubject());
                     return newJws;